final touchups / turn off double click resize
