+++ new-frontend/src/components/MainScreen/MainLayout.js
@@ -2,8 +2,6 @@ import React from "react";
 import SideBar from "./SideBar";
 
 export default function MainLayout() {
-  const remote = require('electron').remote;
-  
   const btnClick = (name) => {
     console.log("clicked", name);
   };
@@ -18,7 +16,7 @@ export default function MainLayout() {
   const sideScreen = {
     backgroundColor: "#ececec",
     flex: 1,
-    height: "100%",
+    height: "100vh",
   };
 
   const cameraScreen = {
@@ -29,37 +27,12 @@ export default function MainLayout() {
     overflow: "hidden",
   };
 
-  // const closeScreen = () => {
-  //   var window = remote.getCurrentWindow();
-  //   window.close(); 
-  // }
-
-  // const minimizeScreen = () => {
-  //   var window = remote.getCurrentWindow();
-  //   window.minimize(); 
-  // }
-
-  // const menuButtonStyle = {
-  //   color: "black",
-  //   backgroundColor: "transparent",
-  //   position: "relative",
-  //   // top: "-45vh",
-  //   // right: "3vh",
-  //   margin: "1vh",
-  //   border: "none",
-  //   cursor: "pointer",
-  //   fontSize: 16,
-  //   fontFamily: "Oxygen",
-  // }
-
   return (
     <div style={flexContainer}>
       <div style={cameraScreen}></div>
       <div style={sideScreen}>
         <SideBar btnClick={btnClick} />
       </div>
-      {/* <button style={menuButtonStyle} onClick={minimizeScreen}>—</button>
-      <button style={menuButtonStyle} onClick={closeScreen}>X</button> */}
     </div>
   );
 }
