+++ new-frontend/src/components/MainScreen/SideBar.js
@@ -1,6 +1,7 @@
 import React from "react";
 import Options from "./Options";
 import CameraOption from "./CameraOption";
+import MenuButtons from "./MenuButtons";
 import { Link } from "react-router-dom";
 import settingsPic from "../../assets/settings.png";
 import aboutPic from "../../assets/about.png";
@@ -8,19 +9,17 @@ import logo from "../../assets/logo.png";
 import cameraPic from "../../assets/webcam.png";
 
 export default function SideBar({ btnClick }) { 
-  const remote = require('electron').remote;
-  
   const logoStyle = {
     height: "10vh",
     position: "relative",
-    top: "-4vh"
+    top: "-4.5vh"
   };
 
   const optionWrapper = {
-    margin: "15vh 2.5vw",
-    height: "60vh",
+    margin: "9vh 2.5vw",
+    height: "67vh",
     display: "flex",
-    flexDirection: "column"
+    flexDirection: "column",
   };
 
   const minorOptions  = {
@@ -38,37 +37,13 @@ export default function SideBar({ btnClick }) {
     flex: 1,
   };
 
-  const closeScreen = () => {
-    var window = remote.getCurrentWindow();
-    window.close(); 
-  }
-
-  const minimizeScreen = () => {
-    var window = remote.getCurrentWindow();
-    window.minimize(); 
-  }
-
-  const menuButtonStyle = {
-    color: "#111111",
-    backgroundColor: "transparent",
-    position: "relative",
-    left: "17.5vw",
-    top: "-7vh",
-    border: "none",
-    cursor: "pointer",
-    fontSize: 15,
-    fontFamily: "Oxygen",
-    fontWeight: "Bold"
-  }
-
   return (
     <div>
       <div style={logoStyle}>
         <img src={logo} alt="logo" width="54.5%" height="auto" />
       </div>
       <div>
-        <button style={menuButtonStyle} onClick={minimizeScreen}>—</button>
-        <button style={menuButtonStyle} onClick={closeScreen}>X</button>
+        <MenuButtons />
       </div>
       <div style={optionWrapper}>
         <div style={minorOptions}>
