+++ new-frontend/src/components/Settings/SideBar.js
@@ -1,5 +1,6 @@
 import React from "react";
 import Options from "../MainScreen/Options";
+import MenuButtons from "../MainScreen/MenuButtons";
 import { Link } from "react-router-dom";
 import homePic from "../../assets/home.png";
 import logo from "../../assets/logo.png";
@@ -8,35 +9,12 @@ export default function SideBar({ btnClick }) {
   const logoStyle = {
     height: "10vh",
     position: "relative",
-    top: "-4vh"
+    top: "-4.5vh"
   };
 
   const optionWrapper = {
-    margin: "17vh 3vw",
+    margin: "12vh 3vw",
   };
-
-  const closeScreen = () => {
-    var window = remote.getCurrentWindow();
-    window.close(); 
-  }
-
-  const minimizeScreen = () => {
-    var window = remote.getCurrentWindow();
-    window.minimize(); 
-  }
-
-  const menuButtonStyle = {
-    color: "#111111",
-    backgroundColor: "transparent",
-    position: "relative",
-    left: "17.5vw",
-    top: "-7vh",
-    border: "none",
-    cursor: "pointer",
-    fontSize: 15,
-    fontFamily: "Oxygen",
-    fontWeight: "Bold"
-  }
   
   return (
     <div>
@@ -44,8 +22,7 @@ export default function SideBar({ btnClick }) {
         <img src={logo} alt="logo" width="54.5%" height="auto" />
       </div>
       <div>
-        <button style={menuButtonStyle} onClick={minimizeScreen}>—</button>
-        <button style={menuButtonStyle} onClick={closeScreen}>X</button>
+        <MenuButtons />
       </div>
       <div style={optionWrapper}>
         <div>
